refactor: move time tests
