+++ cast_test.go
@@ -7,8 +7,6 @@ package cast
 
 import (
 	"encoding/json"
-	"fmt"
-	"path"
 	"testing"
 	"time"
 
@@ -107,17 +105,6 @@ func BenchmarkTooBool(b *testing.B) {
 	}
 }
 
-func BenchmarkCommonTimeLayouts(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		for _, commonLayout := range []string{"2019-04-29", "2017-05-30T00:00:00Z"} {
-			_, err := StringToDateInDefaultLocation(commonLayout, time.UTC)
-			if err != nil {
-				b.Fatal(err)
-			}
-		}
-	}
-}
-
 func TestIndirectPointers(t *testing.T) {
 	c := qt.New(t)
 
@@ -128,295 +115,4 @@ func TestIndirectPointers(t *testing.T) {
 	c.Assert(ToInt(y), qt.Equals, 13)
 	c.Assert(ToInt(z), qt.Equals, 13)
 
-}
-
-func TestToTime(t *testing.T) {
-	c := qt.New(t)
-
-	var jntime, jnetime json.Number
-	_ = json.Unmarshal([]byte("1234567890"), &jntime)
-	_ = json.Unmarshal([]byte("123.4567890"), &jnetime)
-	tests := []struct {
-		input  interface{}
-		expect time.Time
-		iserr  bool
-	}{
-		{"2009-11-10 23:00:00 +0000 UTC", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},   // Time.String()
-		{"Tue Nov 10 23:00:00 2009", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},        // ANSIC
-		{"Tue Nov 10 23:00:00 UTC 2009", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},    // UnixDate
-		{"Tue Nov 10 23:00:00 +0000 2009", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},  // RubyDate
-		{"10 Nov 09 23:00 UTC", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},             // RFC822
-		{"10 Nov 09 23:00 +0000", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},           // RFC822Z
-		{"Tuesday, 10-Nov-09 23:00:00 UTC", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false}, // RFC850
-		{"Tue, 10 Nov 2009 23:00:00 UTC", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},   // RFC1123
-		{"Tue, 10 Nov 2009 23:00:00 +0000", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false}, // RFC1123Z
-		{"2009-11-10T23:00:00Z", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},            // RFC3339
-		{"2018-10-21T23:21:29+0200", time.Date(2018, 10, 21, 21, 21, 29, 0, time.UTC), false},      // RFC3339 without timezone hh:mm colon
-		{"2009-11-10T23:00:00Z", time.Date(2009, 11, 10, 23, 0, 0, 0, time.UTC), false},            // RFC3339Nano
-		{"11:00PM", time.Date(0, 1, 1, 23, 0, 0, 0, time.UTC), false},                              // Kitchen
-		{"Nov 10 23:00:00", time.Date(0, 11, 10, 23, 0, 0, 0, time.UTC), false},                    // Stamp
-		{"Nov 10 23:00:00.000", time.Date(0, 11, 10, 23, 0, 0, 0, time.UTC), false},                // StampMilli
-		{"Nov 10 23:00:00.000000", time.Date(0, 11, 10, 23, 0, 0, 0, time.UTC), false},             // StampMicro
-		{"Nov 10 23:00:00.000000000", time.Date(0, 11, 10, 23, 0, 0, 0, time.UTC), false},          // StampNano
-		{"2016-03-06 15:28:01-00:00", time.Date(2016, 3, 6, 15, 28, 1, 0, time.UTC), false},        // RFC3339 without T
-		{"2016-03-06 15:28:01-0000", time.Date(2016, 3, 6, 15, 28, 1, 0, time.UTC), false},         // RFC3339 without T or timezone hh:mm colon
-		{"2016-03-06 15:28:01", time.Date(2016, 3, 6, 15, 28, 1, 0, time.UTC), false},
-		{"2016-03-06 15:28:01 -0000", time.Date(2016, 3, 6, 15, 28, 1, 0, time.UTC), false},
-		{"2016-03-06 15:28:01 -00:00", time.Date(2016, 3, 6, 15, 28, 1, 0, time.UTC), false},
-		{"2016-03-06 15:28:01 +0900", time.Date(2016, 3, 6, 6, 28, 1, 0, time.UTC), false},
-		{"2016-03-06 15:28:01 +09:00", time.Date(2016, 3, 6, 6, 28, 1, 0, time.UTC), false},
-		{"2006-01-02", time.Date(2006, 1, 2, 0, 0, 0, 0, time.UTC), false},
-		{"02 Jan 2006", time.Date(2006, 1, 2, 0, 0, 0, 0, time.UTC), false},
-		{1472574600, time.Date(2016, 8, 30, 16, 30, 0, 0, time.UTC), false},
-		{int(1482597504), time.Date(2016, 12, 24, 16, 38, 24, 0, time.UTC), false},
-		{int64(1234567890), time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC), false},
-		{int32(1234567890), time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC), false},
-		{uint(1482597504), time.Date(2016, 12, 24, 16, 38, 24, 0, time.UTC), false},
-		{uint64(1234567890), time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC), false},
-		{uint32(1234567890), time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC), false},
-		{jntime, time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC), false},
-		{time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC), time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC), false},
-		// errors
-		{"2006", time.Time{}, true},
-		{jnetime, time.Time{}, true},
-		{testing.T{}, time.Time{}, true},
-	}
-
-	for i, test := range tests {
-		errmsg := qt.Commentf("i = %d", i) // assert helper message
-
-		v, err := ToTimeE(test.input)
-		if test.iserr {
-			c.Assert(err, qt.IsNotNil)
-			continue
-		}
-
-		c.Assert(err, qt.IsNil)
-		c.Assert(v.UTC(), qt.Equals, test.expect, errmsg)
-
-		// Non-E test
-		v = ToTime(test.input)
-		c.Assert(v.UTC(), qt.Equals, test.expect, errmsg)
-	}
-}
-
-func TestToDurationE(t *testing.T) {
-	c := qt.New(t)
-
-	var td time.Duration = 5
-	var jn json.Number
-	_ = json.Unmarshal([]byte("5"), &jn)
-
-	tests := []struct {
-		input  interface{}
-		expect time.Duration
-		iserr  bool
-	}{
-		{time.Duration(5), td, false},
-		{int(5), td, false},
-		{int64(5), td, false},
-		{int32(5), td, false},
-		{int16(5), td, false},
-		{int8(5), td, false},
-		{uint(5), td, false},
-		{uint64(5), td, false},
-		{uint32(5), td, false},
-		{uint16(5), td, false},
-		{uint8(5), td, false},
-		{float64(5), td, false},
-		{float32(5), td, false},
-		{jn, td, false},
-		{string("5"), td, false},
-		{string("5ns"), td, false},
-		{string("5us"), time.Microsecond * td, false},
-		{string("5µs"), time.Microsecond * td, false},
-		{string("5ms"), time.Millisecond * td, false},
-		{string("5s"), time.Second * td, false},
-		{string("5m"), time.Minute * td, false},
-		{string("5h"), time.Hour * td, false},
-		// errors
-		{"test", 0, true},
-		{testing.T{}, 0, true},
-	}
-
-	for i, test := range tests {
-		errmsg := qt.Commentf("i = %d", i) // assert helper message
-
-		v, err := ToDurationE(test.input)
-		if test.iserr {
-			c.Assert(err, qt.IsNotNil)
-			continue
-		}
-
-		c.Assert(err, qt.IsNil)
-		c.Assert(v, qt.Equals, test.expect, errmsg)
-
-		// Non-E test
-		v = ToDuration(test.input)
-		c.Assert(v, qt.Equals, test.expect, errmsg)
-	}
-}
-
-func TestToTimeWithTimezones(t *testing.T) {
-	c := qt.New(t)
-
-	est, err := time.LoadLocation("EST")
-	c.Assert(err, qt.IsNil)
-
-	irn, err := time.LoadLocation("Iran")
-	c.Assert(err, qt.IsNil)
-
-	swd, err := time.LoadLocation("Europe/Stockholm")
-	c.Assert(err, qt.IsNil)
-
-	// Test same local time in different timezones
-	utc2016 := time.Date(2016, time.January, 1, 0, 0, 0, 0, time.UTC)
-	est2016 := time.Date(2016, time.January, 1, 0, 0, 0, 0, est)
-	irn2016 := time.Date(2016, time.January, 1, 0, 0, 0, 0, irn)
-	swd2016 := time.Date(2016, time.January, 1, 0, 0, 0, 0, swd)
-	loc2016 := time.Date(2016, time.January, 1, 0, 0, 0, 0, time.Local)
-
-	for i, format := range timeFormats {
-		format := format
-		if format.typ == timeFormatTimeOnly {
-			continue
-		}
-
-		nameBase := fmt.Sprintf("%d;timeFormatType=%d;%s", i, format.typ, format.format)
-
-		t.Run(path.Join(nameBase), func(t *testing.T) {
-			est2016str := est2016.Format(format.format)
-			swd2016str := swd2016.Format(format.format)
-
-			t.Run("without default location", func(t *testing.T) {
-				c := qt.New(t)
-				converted, err := ToTimeE(est2016str)
-				c.Assert(err, qt.IsNil)
-				if format.hasTimezone() {
-					// Converting inputs with a timezone should preserve it
-					assertTimeEqual(t, est2016, converted)
-					assertLocationEqual(t, est, converted.Location())
-				} else {
-					// Converting inputs without a timezone should be interpreted
-					// as a local time in UTC.
-					assertTimeEqual(t, utc2016, converted)
-					assertLocationEqual(t, time.UTC, converted.Location())
-				}
-			})
-
-			t.Run("local timezone without a default location", func(t *testing.T) {
-				c := qt.New(t)
-				converted, err := ToTimeE(swd2016str)
-				c.Assert(err, qt.IsNil)
-				if format.hasTimezone() {
-					// Converting inputs with a timezone should preserve it
-					assertTimeEqual(t, swd2016, converted)
-					assertLocationEqual(t, swd, converted.Location())
-				} else {
-					// Converting inputs without a timezone should be interpreted
-					// as a local time in UTC.
-					assertTimeEqual(t, utc2016, converted)
-					assertLocationEqual(t, time.UTC, converted.Location())
-				}
-			})
-
-			t.Run("nil default location", func(t *testing.T) {
-				c := qt.New(t)
-
-				converted, err := ToTimeInDefaultLocationE(est2016str, nil)
-				c.Assert(err, qt.IsNil)
-				if format.hasTimezone() {
-					// Converting inputs with a timezone should preserve it
-					assertTimeEqual(t, est2016, converted)
-					assertLocationEqual(t, est, converted.Location())
-				} else {
-					// Converting inputs without a timezone should be interpreted
-					// as a local time in the local timezone.
-					assertTimeEqual(t, loc2016, converted)
-					assertLocationEqual(t, time.Local, converted.Location())
-				}
-
-			})
-
-			t.Run("default location not UTC", func(t *testing.T) {
-				c := qt.New(t)
-
-				converted, err := ToTimeInDefaultLocationE(est2016str, irn)
-				c.Assert(err, qt.IsNil)
-				if format.hasTimezone() {
-					// Converting inputs with a timezone should preserve it
-					assertTimeEqual(t, est2016, converted)
-					assertLocationEqual(t, est, converted.Location())
-				} else {
-					// Converting inputs without a timezone should be interpreted
-					// as a local time in the given location.
-					assertTimeEqual(t, irn2016, converted)
-					assertLocationEqual(t, irn, converted.Location())
-				}
-
-			})
-
-			t.Run("time in the local timezone default location not UTC", func(t *testing.T) {
-				c := qt.New(t)
-
-				converted, err := ToTimeInDefaultLocationE(swd2016str, irn)
-				c.Assert(err, qt.IsNil)
-				if format.hasTimezone() {
-					// Converting inputs with a timezone should preserve it
-					assertTimeEqual(t, swd2016, converted)
-					assertLocationEqual(t, swd, converted.Location())
-				} else {
-					// Converting inputs without a timezone should be interpreted
-					// as a local time in the given location.
-					assertTimeEqual(t, irn2016, converted)
-					assertLocationEqual(t, irn, converted.Location())
-				}
-
-			})
-
-		})
-
-	}
-}
-
-func assertTimeEqual(t *testing.T, expected, actual time.Time) {
-	t.Helper()
-	// Compare the dates using a numeric zone as there are cases where
-	// time.Parse will assign a dummy location.
-	qt.Assert(t, actual.Format(time.RFC1123Z), qt.Equals, expected.Format(time.RFC1123Z))
-}
-
-func assertLocationEqual(t *testing.T, expected, actual *time.Location) {
-	t.Helper()
-	qt.Assert(t, locationEqual(expected, actual), qt.IsTrue)
-}
-
-func locationEqual(a, b *time.Location) bool {
-	// A note about comparring time.Locations:
-	//   - can't only compare pointers
-	//   - can't compare loc.String() because locations with the same
-	//     name can have different offsets
-	//   - can't use reflect.DeepEqual because time.Location has internal
-	//     caches
-
-	if a == b {
-		return true
-	} else if a == nil || b == nil {
-		return false
-	}
-
-	// Check if they're equal by parsing times with a format that doesn't
-	// include a timezone, which will interpret it as being a local time in
-	// the given zone, and comparing the resulting local times.
-	tA, err := time.ParseInLocation("2006-01-02", "2016-01-01", a)
-	if err != nil {
-		return false
-	}
-
-	tB, err := time.ParseInLocation("2006-01-02", "2016-01-01", b)
-	if err != nil {
-		return false
-	}
-
-	return tA.Equal(tB)
 }
